@@ -4,7 +4,9 @@
     using Microsoft.AspNetCore.Hosting;
     using Microsoft.Extensions.Configuration;
     using Microsoft.Extensions.DependencyInjection;
+    using Microsoft.Extensions.Hosting;
     using TaskManager.Tests.Web.Infrastructure.Hubs;
+    using TaskManager.Tests.WebFunctional.Extensions;
 
     public class TestStartup
     {
@@ -20,17 +22,23 @@
             services.InjectServices();
         }
 
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
             }
 
-            app.UseAuthentication();
-            app.UseMvc();
             app.UseHttpsRedirection();
-            app.UseSignalR(routes => routes.MapHub<NotificationTestHub>("/Notifications"));
+            app.UseRouting();
+            app.UseCors(builder => builder.AllowAnyHeader().AllowAnyHeader().AllowAnyOrigin());
+            app.UseAuthentication();
+            app.UseAuthorization();
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapControllers();
+                endpoints.MapHub<NotificationTestHub>("/Notifications");
+            });
         }
     }
 }
